refactor(tests): 正交拆分 kg_entity_service + graph_service 单体测试文件
1. test_kg_entity_service_unit.py (876 行) → entity/ 子目录 6 文件：
   按 4 个 Test 类正交分解为 test_sync_entity / test_sync_relation /
   test_batch_sync / test_get_top_entities + conftest；
2. test_graph_service.py (840 行) → graph/ 子目录 5 文件：
   按 CRUD / build+cancel / result types 三个正交维度分解，
   共享 Fake Repository 和 patch helper 收敛至 conftest；
3. 全量回归 1,661 测试通过，无破坏性变更。
